requirements.txt
asgiref==3.4.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
